separate search and card component, added favorites to heart icon
