+++ src/components/CSS/Search.css
@@ -54,6 +54,6 @@
   color: #fff;
 }
 
-.turquoise  {
+.favorite  {
   color: rgb(3, 200, 200);
 }
